sprite.go: Move sprite struct to own file
